+++ encoder/object_encoder.go
@@ -145,7 +145,7 @@ func (e *objectEncoder) AppendByteString(v []byte) {
 
 func (e *objectEncoder) AppendComplex128(v complex128) {
 	// Cast to a platform-independent, fixed-size type.
-	r, i := float64(real(v)), float64(imag(v))
+	r, i := real(v), imag(v)
 	e.buf.AppendByte('"')
 	// Because we're always in a quoted string, we can use strconv without
 	// special-casing NaN and +/-Inf.
